requirements.txt
Flask
pymongo
google-generativeai
